修改天数支持小数   验证整数正则修改

+++ thai-crm-master/src/main/java/com/magic/thai/db/domain/Goods.java
@@ -63,7 +63,7 @@ public class Goods {
 	private String hotelName;
 
 	@Column(name = "travel_days")
-	private int travelDays = 1;// 行程天数
+	private double travelDays = 1d;// 行程天数
 
 	@Expose
 	@Column(name = "goods_count")
@@ -181,11 +181,11 @@ public class Goods {
 		this.hotelId = hotelId;
 	}
 
-	public int getTravelDays() {
+	public double getTravelDays() {
 		return travelDays;
 	}
 
-	public void setTravelDays(int travelDays) {
+	public void setTravelDays(double travelDays) {
 		this.travelDays = travelDays;
 	}
